refactor: eventlistener에 있던 로직을 usecase, facade로 책임 분리 - #223
